Move platform and roster state factories
