+++ platform-sdk/consensus-platformstate/src/test/java/org/hiero/consensus/platformstate/PlatformStateUtilsTest.java
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.hiero.consensus.platformstate;
 
-import static com.swirlds.platform.test.fixtures.PlatformStateUtils.randomPlatformState;
+import static org.hiero.consensus.platformstate.PlatformStateFactory.randomPlatformState;
 import static com.swirlds.state.test.fixtures.merkle.VirtualMapStateTestUtils.createTestState;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.hiero.base.utility.test.fixtures.RandomUtils.nextLong;
