+++ platform-sdk/swirlds-platform-core/src/testFixtures/java/com/swirlds/platform/test/fixtures/PlatformTestUtils.java
@@ -1,10 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.swirlds.platform.test.fixtures;
 
-import com.hedera.hapi.node.base.ServiceEndpoint;
-import com.hedera.hapi.node.state.roster.Roster;
-import com.hedera.hapi.node.state.roster.RosterEntry;
-import com.hedera.pbj.runtime.io.buffer.Bytes;
 import com.swirlds.common.context.PlatformContext;
 import com.swirlds.common.test.fixtures.platform.TestPlatformContextBuilder;
 import com.swirlds.config.extensions.test.fixtures.TestConfigBuilder;
@@ -14,15 +10,7 @@ import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.security.cert.CertificateEncodingException;
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
 import java.util.function.Function;
-import org.hiero.consensus.model.node.KeysAndCerts;
-import org.hiero.consensus.model.node.NodeId;
 
 /**
  * Platform level unit test base class for common setup and teardown.
@@ -81,36 +69,4 @@ public class PlatformTestUtils {
         return platformContextBuilder.build();
     }
 
-    /**
-     * Create a Roster for the given signers
-     */
-    @NonNull
-    public static Roster generateRoster(@NonNull final Map<NodeId, KeysAndCerts> signers) {
-        final List<RosterEntry> rosterEntries = new ArrayList<>();
-        for (final Entry<NodeId, KeysAndCerts> signer : signers.entrySet()) {
-            rosterEntries.add(createRosterEntry(signer.getKey(), signer.getValue()));
-        }
-        rosterEntries.sort(Comparator.comparingLong(RosterEntry::nodeId));
-        return Roster.newBuilder().rosterEntries(rosterEntries).build();
-    }
-
-    @NonNull
-    private static RosterEntry createRosterEntry(
-            @NonNull final NodeId nodeId, @NonNull final KeysAndCerts keysAndCerts) {
-        try {
-            final long id = nodeId.id();
-            final byte[] certificate = keysAndCerts.sigCert().getEncoded();
-            return RosterEntry.newBuilder()
-                    .nodeId(id)
-                    .weight(500)
-                    .gossipCaCertificate(Bytes.wrap(certificate))
-                    .gossipEndpoint(ServiceEndpoint.newBuilder()
-                            .domainName(String.format("node-%d", id))
-                            .port(8082)
-                            .build())
-                    .build();
-        } catch (final CertificateEncodingException e) {
-            throw new RuntimeException("Exception while creating roster entry", e);
-        }
-    }
 }
